services/notifier/requirements.txt
nats-py>=2.9.0
requests>=2.31.0
